Fix: Resolve focus loss in cost editor
Extracts `MobileCostItem` and `DesktopCostItem` components from `WorkOrderCostsEditor` to prevent re-renders and focus loss on input changes. Wraps extracted components with `React.memo` for optimization and ensures stable prop references.
