+++ src/components/work-orders/WorkOrderCostsEditor.tsx
@@ -1,10 +1,11 @@
 
-import React from 'react';
+import React, { useCallback } from 'react';
 import { Button } from '@/components/ui/button';
-import { Input } from '@/components/ui/input';
-import { Trash2, Plus } from 'lucide-react';
+import { Plus } from 'lucide-react';
 import { WorkOrderCostItem } from '@/hooks/useWorkOrderCostsState';
 import { useIsMobile } from '@/hooks/use-mobile';
+import MobileCostItem from './MobileCostItem';
+import DesktopCostItem from './DesktopCostItem';
 
 interface WorkOrderCostsEditorProps {
   costs: WorkOrderCostItem[];
@@ -23,133 +24,19 @@ const WorkOrderCostsEditor: React.FC<WorkOrderCostsEditorProps> = ({
 }) => {
   const isMobile = useIsMobile();
   
-  const formatCurrency = (cents: number) => {
+  const formatCurrency = useCallback((cents: number) => {
     return new Intl.NumberFormat('en-US', {
       style: 'currency',
       currency: 'USD',
       minimumFractionDigits: 2
     }).format(cents / 100);
-  };
+  }, []);
 
-  const calculateSubtotal = () => {
+  const calculateSubtotal = useCallback(() => {
     return costs.reduce((sum, cost) => sum + cost.total_price_cents, 0);
-  };
+  }, [costs]);
 
-  const MobileCostItem = ({ cost }: { cost: WorkOrderCostItem }) => (
-    <div className="bg-muted/50 rounded-lg p-4 space-y-3">
-      <div className="flex items-center justify-between">
-        <span className="text-sm font-medium text-muted-foreground">Description</span>
-        <Button
-          type="button"
-          variant="ghost"
-          size="sm"
-          onClick={() => onRemoveCost(cost.id)}
-          className="text-red-600 hover:text-red-700 h-6 w-6 p-0"
-          disabled={costs.length === 1}
-        >
-          <Trash2 className="h-4 w-4" />
-        </Button>
-      </div>
-      
-      <Input
-        value={cost.description}
-        onChange={(e) => onUpdateCost(cost.id, 'description', e.target.value)}
-        placeholder="Enter description..."
-        className="h-9"
-      />
-      
-      <div className="grid grid-cols-2 gap-3">
-        <div>
-          <label className="text-sm font-medium text-muted-foreground block mb-1">
-            Quantity
-          </label>
-          <Input
-            type="number"
-            step="0.01"
-            min="0.01"
-            value={cost.quantity}
-            onChange={(e) => onUpdateCost(cost.id, 'quantity', parseFloat(e.target.value) || 1)}
-            placeholder="Qty"
-            className="h-9"
-          />
-        </div>
-        
-        <div>
-          <label className="text-sm font-medium text-muted-foreground block mb-1">
-            Unit Price
-          </label>
-          <div className="flex items-center gap-1">
-            <span className="text-sm text-muted-foreground">$</span>
-            <Input
-              type="number"
-              step="0.01"
-              min="0"
-              value={cost.unit_price_cents / 100}
-              onChange={(e) => onUpdateCost(cost.id, 'unit_price_cents', Math.round((parseFloat(e.target.value) || 0) * 100))}
-              placeholder="0.00"
-              className="h-9"
-            />
-          </div>
-        </div>
-      </div>
-      
-      <div className="flex items-center justify-between pt-2 border-t">
-        <span className="text-sm font-medium text-muted-foreground">Total:</span>
-        <span className="font-semibold text-lg">
-          {formatCurrency(cost.total_price_cents)}
-        </span>
-      </div>
-    </div>
-  );
-
-  const DesktopCostItem = ({ cost }: { cost: WorkOrderCostItem }) => (
-    <div className="flex items-center justify-between p-3 bg-muted/50 rounded-lg">
-      <div className="flex-1 grid grid-cols-4 gap-4 items-center">
-        <Input
-          value={cost.description}
-          onChange={(e) => onUpdateCost(cost.id, 'description', e.target.value)}
-          placeholder="Enter description..."
-          className="h-8"
-        />
-        <Input
-          type="number"
-          step="0.01"
-          min="0.01"
-          value={cost.quantity}
-          onChange={(e) => onUpdateCost(cost.id, 'quantity', parseFloat(e.target.value) || 1)}
-          placeholder="Qty"
-          className="h-8"
-        />
-        <div className="flex items-center gap-1">
-          <span className="text-sm text-muted-foreground">$</span>
-          <Input
-            type="number"
-            step="0.01"
-            min="0"
-            value={cost.unit_price_cents / 100}
-            onChange={(e) => onUpdateCost(cost.id, 'unit_price_cents', Math.round((parseFloat(e.target.value) || 0) * 100))}
-            placeholder="0.00"
-            className="h-8"
-          />
-        </div>
-        <div className="flex items-center justify-between">
-          <span className="font-semibold text-sm">
-            {formatCurrency(cost.total_price_cents)}
-          </span>
-          <Button
-            type="button"
-            variant="ghost"
-            size="sm"
-            onClick={() => onRemoveCost(cost.id)}
-            className="text-red-600 hover:text-red-700"
-            disabled={costs.length === 1}
-          >
-            <Trash2 className="h-4 w-4" />
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
+  const canRemove = costs.length > 1;
 
   return (
     <div className="space-y-4">
@@ -191,9 +78,19 @@ const WorkOrderCostsEditor: React.FC<WorkOrderCostsEditorProps> = ({
             {costs.map((cost) => (
               <div key={cost.id}>
                 {isMobile ? (
-                  <MobileCostItem cost={cost} />
+                  <MobileCostItem 
+                    cost={cost}
+                    onRemoveCost={onRemoveCost}
+                    onUpdateCost={onUpdateCost}
+                    canRemove={canRemove}
+                  />
                 ) : (
-                  <DesktopCostItem cost={cost} />
+                  <DesktopCostItem 
+                    cost={cost}
+                    onRemoveCost={onRemoveCost}
+                    onUpdateCost={onUpdateCost}
+                    canRemove={canRemove}
+                  />
                 )}
               </div>
             ))}
